Restructured the Packaging Structure of Flask Project
