--- Main/Backend/Config.py
@@ -1,15 +1,0 @@
-import os
-
-UPLOAD_FOLDER = os.path.abspath(os.path.dirname(__file__))+'/resources/VoiceUploads/'
-ALLOWED_EXTENSIONS = {'txt', 'mp3','wav'}
-
-DEV_IP = 'localhost'
-STG_IP = ''
-PROD_IP = ''
-
-DB_DEV_IP = 'mongodb://localhost:27017/'
-DB_STG_IP = ''
-DB_PROD_IP = ''
-
-DB_SCHEMA = 'Phoenix-Hospital'
-IMIST_AMBO_TEMPLATE = 'imist_ambo_category'
