--- Main/Backend/Model.py
@@ -1,186 +1,0 @@
-from Config import *
-from pymongo import MongoClient
-
-client = MongoClient(DB_DEV_IP)
-mydb = client[DB_SCHEMA]
-
-class IMIST_AMBO:
-
-	def __init__(self):
-		self.mytable = mydb[IMIST_AMBO_TEMPLATE]
-		
-		self.map_keyword_category = [
-			{"keyword": "age",
-			 "category": "identification"},
-			{"keyword": "male",
-			 "category": "identification"},
-			{"keyword": "female",
-			 "category": "identification"},
-			{"keyword": "mca",
-			 "category": "mechanism"},
-			{"keyword": "rollov",
-			 "category": "mechanism"},
-			{"keyword": "eject",
-			 "category": "mechanism"},
-			{"keyword": "death other occupant",
-			 "category": "mechanism"},
-			{"keyword": "pedestrian",
-			 "category": "mechanism"},
-			{"keyword": "motorcyclist",
-			 "category": "mechanism"},
-			{"keyword": "cyclist",
-			 "category": "mechanism"},
-			 {"keyword": "motorcycl",
-			 "category": "mechanism"},
-			{"keyword": "cycl",
-			 "category": "mechanism"},
-			{"keyword": "fall",
-			 "category": "mechanism"},
-			{"keyword": "burn",
-			 "category": "mechanism"},
-			{"keyword": "explos",
-			 "category": "mechanism"},
-			{"keyword": "trap",
-			 "category": "mechanism"},
-			{"keyword": "time entrap",
-			 "category": "mechanism"},
-			{"keyword": "mba",
-			 "category": "mechanism"},
-			{"keyword": "extric",
-			 "category": "mechanism"},
-			{"keyword": "fatal",
-			 "category": "mechanism"},
-			{"keyword": "eject",
-			 "category": "mechanism"},
-			{"keyword": "penetr",
-			 "category": "injury"},
-			{"keyword": "blunt",
-			 "category": "injury"},
-			{"keyword": "head",
-			 "category": "injury"},
-			{"keyword": "neck",
-			 "category": "injury"},
-			{"keyword": "chest",
-			 "category": "injury"},
-			{"keyword": "abdomen",
-			 "category": "injury"},
-			{"keyword": "pelvi",
-			 "category": "injury"},
-			{"keyword": "axilla",
-			 "category": "injury"},
-			{"keyword": "groin",
-			 "category": "injury"},
-			{"keyword": "limb",
-			 "category": "injury"},
-			{"keyword": "amput",
-			 "category": "injury"},
-			{"keyword": "crush",
-			 "category": "injury"},
-			{"keyword": "spinal",
-			 "category": "injury"},
-			{"keyword": "tension pneumothorax",
-			 "category": "injury"},
-			{"keyword": "rigid abdomen",
-			 "category": "injury"},
-			{"keyword": "fractur",
-			 "category": "injury"},
-			{"keyword": "pelvi",
-			 "category": "injury"},
-			{"keyword": "facial burn",
-			 "category": "injury"},
-			{"keyword": "disloc",
-			 "category": "injury"},
-			{"keyword": "eviscer",
-			 "category": "injury"},
-			{"keyword": "blast",
-			 "category": "injury"},    
-			{"keyword": "pr",
-			 "category": "signs"},
-			{"keyword": "bp",
-			 "category": "signs"},
-			{"keyword": "gcs",
-			 "category": "signs"},
-			{"keyword": "evm",
-			 "category": "signs"},
-			{"keyword": "pupil size",
-			 "category": "signs"},
-			{"keyword": "reactiv",
-			 "category": "signs"},
-			{"keyword": "rr",
-			 "category": "signs"},
-			{"keyword": "t degree",
-			 "category": "signs"},
-			{"keyword": "spo 2",
-			 "category": "signs"},
-			{"keyword": "sob",
-			 "category": "signs"},
-			{"keyword": "sbp",
-			 "category": "signs"},
-			{"keyword": "cervic collar",
-			 "category": "treatment"},
-			{"keyword": "op airway",
-			 "category": "treatment"},
-			{"keyword": "np airway",
-			 "category": "treatment"},
-			{"keyword": "lma",
-			 "category": "treatment"},
-			{"keyword": "ett",
-			 "category": "treatment"},
-			{"keyword": "rsi",
-			 "category": "treatment"},
-			{"keyword": "ventil",
-			 "category": "treatment"},
-			{"keyword": "chest decompress",
-			 "category": "treatment"},
-			{"keyword": "iv access",
-			 "category": "treatment"},
-			{"keyword": "iv hartmann",
-			 "category": "treatment"},
-			{"keyword": "methoxyfluran",
-			 "category": "treatment"},
-			{"keyword": "maxolon",
-			 "category": "treatment"},
-			{"keyword": "morphin",
-			 "category": "treatment"},
-			{"keyword": "midazolam",
-			 "category": "treatment"},
-			{"keyword": "fentanyl",
-			 "category": "treatment"},
-			{"keyword": "suxamethonium",
-			 "category": "treatment"},
-			{"keyword": "pancuronium",
-			 "category": "treatment"},
-			{"keyword": "adrenalin",
-			 "category": "treatment"},
-			{"keyword": "intub",
-			 "category": "treatment"},
-			{"keyword": "haemost dressing",
-			 "category": "treatment"},
-			{"keyword": "toumiquet",
-			 "category": "treatment"},
-			{"keyword": "blood transfus",
-			 "category": "treatment"},
-			{"keyword": "neuromuscular block",
-			 "category": "treatment"},
-			{"keyword": "allergi",
-			 "category": "other"},
-			{"keyword": "mass casualti",
-			 "category": "other"},
-			{"keyword": "inter-hospit transfer",
-			 "category": "other"},
-			{"keyword": "pregnanc",
-			 "category": "other"},
-			{"keyword": "pregnant",
-			 "category": "other"},
-			{"keyword": "co-morbid",
-			 "category": "other"},
-			{"keyword": "anticoagul therapi",
-			 "category": "other"}
-			]
-
-	def insert(self):
-		self.mytable.insert(self.map_keyword_category)
-
-if __name__=='__main__':
-	imist_ambo = IMIST_AMBO()
-	imist_ambo.insert()
