+++ Main/Backend/P5VoiceToText/uploadOrSelectFiles/utils.py
@@ -1,0 +1,7 @@
+from P5VoiceToText.config import Config
+
+allowed_extensions = Config.ALLOWED_EXTENSIONS
+
+def allowed_file(filename):
+    return '.' in filename and \
+           filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
