--- Main/Backend/requirements.txt
@@ -1,3 +1,0 @@
-flask
-flask_cors
-pymongo-3.10.1
