--- Main/Backend/resources/VoiceUploads/why_this_role.txt
@@ -1,1 +1,0 @@
-This role is fascinating as it requires most of the technologies that I have previously worked and some of the technologies which I wanted to explore, which is a balance between becoming  an asset to organization and learning new technologies
