--- Main/Backend/some.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body>
-
-<h1>The input accept attribute</h1>
-
-<form action="/action_page.php">
-  <label for="img">Select image:</label>
-  <input type="file" id="img" name="img" accept="image/*">
-  <input type="submit">
-</form>
-
-<p><strong>Note:</strong> Because of security issues, this example will not allow you to upload files.</p>
-
-</body>
-</html>
