--- Main/Backend/testfile.txt
@@ -1,1 +1,0 @@
-Hi how are you I'm good how are you
